Configure Fargate Spot tasks (#10)
To reduce costs we setup ECS to deploy Fargate Spot tasks
along with on demand task.  The idea is to ensure that there
is at least one fargate on demand task for stability while
all other tasks are run with spot tasks for cost efficiency.

AWS cannot do an in place update of this change.
The update requires stacks to be deleted and redeployed.